user feedback during search for broken symlinks
If you're using less, partial lines don't print. This means the
"searching for broken links..." message isn't printed until the search
is complete and the line is finished with either "OK" or "found broken
links".

This patch splits the searching and results messages so that they're on
different lines and therefore do the right thing when piped to less (or
another pager).